Initial Tray Item Support
I added in the initial Tray item support using the tray-item crate.
Right now the code is mostly straight out of the Windows example (as
that is what I was compiling on) and is Windows specific. This commit
does not currently build on Linux. I'll need to come back and and do
some clean up to get this working on Linux as well.

The *.ico files are all straight out of the Tray-item crate Windows
example, I'll need to remove them later and put in my own icons.

Tomato.svg is a possible new logo I'm toying with, either maybe as the
tray icon, or as a proper replacement for the quickly hacked together
logo based on the pomotroid logo