Adding in What should be basic Linux Tray Icon Support
This now compiles on Linux and I believe in theory works, however the
tray icon isn't showing up on my VM. I'm running on an Arch VM with
Gnome and a Tray Icon Extension. Not sure if the problem is the
extension or the program. I'm committing the code now, but I need to
come back and do a bit more testing.

I've also updated to a new logo, and swapped out the tray icons for a
version of that same logo.