Some preliminary Changing logo Color
Some early stage logic for changing the logo colors with the theme.
Still need to be able to read in the theme json, update the model, read
the values from the model to update the SVG etc.

Fixed the slide over animation for the tab container.

Adding the actual theme json files from the Pomotroid project into the
repo