Basic loading of themes from Json
This commit has basic support for loading themes in the Pomotroid Json
format from a themes directory under the executable folder. It's a bit
crude at the moment, but it gets the job done.

Need to add support for proper config file locations such as the XDG
config directory spec etc.