Add in basic support for swapping the logo colors with the theme
Added basic support to update the colors of the logo on the about screen
to match the theme. this closes Issue #11.

This also might close issue #19 (which really should have been part of
Issue #11), but I'm not sure yet if I want to try and color change the
app icon as well as the about screen logo....