Preliminary work on vertical slider and Volume slider popup
This is some early work to get the existing slider to work in a
horizontal or vertical fashion, so I could use it for the volume slider.

The volume slider in Pomotroid also has a slightly different behavior
with colors than the timer config sliders. It actually has a hover color
for the grabber, and the "filled" part of the track is the same color
as the empty part of the track. So some changes were made to give the
slider a bit more configurability there.