Added a volume level
Added a new volume level value and tied it to the new slider. Set the
mute/muted icon to actually change when the volume is 0. Noticed some
inconsistencies in the icon svg file dimensions and corrected them.

I'll still need to make some tweaks here, in Pomotroid you hover over
mute icon to get the volume slider, and if you click the icon it just
mutes it...which I like and makes sense. the touch area has the clicked
callback which right now is being used to show the popup with the volume
slider, but hover has no callback it's just a state. Need to figure out
the best way to get similar behavior.