Added in initial work for changing between timers
This work allows changing between the timers at the appropriate time.
Changing from Focus to Short Break, and then repeating based on the
number of rounds, before going to a Long Break.

This also now updates the text based on the active timer, and the
current round number, plus set number of rounds actually display
correctly instead of being a static string.