Early work support for saving the preferences
I'm not entirely sure I like how this functions, it seems a bit crude
There is also a very strange bug I'm having with saving the preferences
json file (documented in Issue #32). However the basics of tying stuff
together and saving out some of the settings are in place.

I also upgraded the dependencies to see if that would make a difference
for the strange but I was getting.