Changing the word separators in Slint identifiers
I typically use _ as a seperator in variables and identifiers.
It seems in Slint using - is preferred (as I understand it comes from
CSS). I had a bit of a mix of both all throughout, I've just cleaned
this up so it should all be using - as the separator now