Cleaned up Theme loading folder location
I forgot that as a temporary to get theme loading working I had it look
for a themes subdirectory off the executable's local dir. I need to move
the theme loading into the settings module, but as a quick patch to get
it looking in the correct directory I've made the get_dir from the
settings module public, and am using that.