Initial Global Hot Key Support
There is a bit of restructuring here, as well as basic support for
global hot keys. I have only tested it on Windows 11. It should in
theory work on most "modern" versions of Windows, and Linux with X11, as
well as possibly MacOS. I don't have a Mac to test, and I haven't tested
it on X11 yet.

Wayland support is waiting on this bug to be resolved before it will work:
https://github.com/tauri-apps/global-hotkey/issues/28