Improved the Theme loading
Improved the Theme loading by moving it to a global property. This
closes out Issue #26, and resolves the issue I found while working on
adding in the skip function where it would call load_themes on every
tick of the timer.

This also now adds the skip functionality, both using the global hot key
(still not configurable) and using the button the main screen.