Added basic notification support
The notification support is very basic, there is no code to use any of
the extra functionality between Windows/Linux/Mac that is available. But
it's in place and working.

I also fixed a bug where I wasn't handling the between timers correctly.
At the end of a round instead of doing a Long Break, it would do a Short
Break and then the Long Break. This has now been corrected.