Add support for Always on Top
Added support for Always on Top and disabling Always on Top during
breaks. This mostly resolves Issue #67 and #68. However this doesn't
work on Wayland, as I understand it's not currently supported.

I probably need to come up with a way to either hide or disable the
settings for things that don't function under Wayland. Until I decide
on the way forward with that, I'll leave the above two issues open.

I also probably need to create a table in the README with things that
don't work under Wayland. The two off the top of my head are this, and
Global Hot Keys.