Adds support for disabling things under Wayland
I've added some code to try and determine if the program is running
under Wayland. If it is certain features in the settings are disabled.
Right now that is the "Always On Top" settings, and the Global Hot Keys.

I'm not sure what is involved in getting Always On Top working under
Wayland, as I haven't found a whole lot of information on it yet. So for
now I'm using this to disable those settings, and will be closing Issue
#67 and Issue #68.

I'm also disabling the inputs for the Global Hot Keys, However as there
are still other issues with Global Hot Keys, and there is at least a
tracking issue for adding Wayland support to the Global Hot Key crate
using org.freedesktop.portal.GlobalShortcuts (although I don't think
anyone has started working on it) I'll leave the Issued for GHKs open.