Setup and use the filter model on the configs
This uses the filter model to hide settings that are not currently
active, such as things not currently supported on Wayland, or the always
on top on break (which is only shown when always on top is true).

There is a delay to allow the slide out animation on he always on top on
break, however right now it's not working. I got the animation working
in a sample project, but it's not working here and I'm not sure what
I've overlooked. Since the filter model is working though, I'm commiting
this.