The Animation on the Config page now work
I managed to get the Always On Top on Break to slide in and slide out
when the Always on top options is altered. This now matches how it looks
in Pomotroid. (or close enough to it, I don't if the timing on my
animation is exactly the same, but I'm happy with how it looks)

This closes out Issue #75