Mostly working Slint Progress Circle
Mainly due to a bug with rending line caps in Slint:
https://github.com/slint-ui/slint/issues/8165

I've reverted back to using a cleaned up version of the older
path logic that drew the shape inside an out a solid object.
It actually seems to work better, because the animation on the
progress was causing some strange quirks on the single line path
when it switched from the big arc to the small arc.

There does seem to be a small issue where right now the
animation on the progress tick is being ignore though. It seems
to work fine when using the Slint "Show Preview" to demo the
progress circle, but when running the app, it doesn't seem to be
animating.