Moved the Settings on_theme_changed closure
The closure for the on_theme_changed callback was moved out of the
middle of the main() routine. The main logic
of this has been converted into a function inside the settings
module, which gets called from the main() routine.
Since the color_to_hex_string function and LOGO_BYTES logo svg
were only used inside this closure, it also made sense to move them to
the settings module